sign liquid pset on shared util function
